refactor(toolbar): refactored toolbar buttons to a react component (ToolbarButton) to improve extensibility
