+++ paint-frontend/src/App.jsx
@@ -13,16 +13,16 @@ function App() {
       <Toolbar
         selectedTool={selectedShape}
         setSelectedTool={setSelectedShape}
-        primaryColor={firstColor}
-        setPrimaryColor={setFirstColor}
-        secondaryColor={secondColor}
-        setSecondaryColor={setSecondColor}
+        fillColor={firstColor}
+        setFillColor={setFirstColor}
+        strokeColor={secondColor}
+        setStrokeColor={setSecondColor}
       />
       <Canvas
         selectedShape={selectedShape}
         firstColor={firstColor}
         secondColor={secondColor}
-      />
+      ></Canvas>
     </div>
   );
 }
