+++ paint-frontend/src/styles/Toolbar.css
@@ -3,7 +3,7 @@
   align-items: center;
   justify-content: space-around;
   height: 35px;
-  width: 500px;
+  width: 100vw;
   background-color: #f0f0f0;
   border: 1px solid #ccc;
   border-radius: 5px;
